Data Exists

this function is now deprecated as of version 0.0.0.9045

sbf_data_exists(x_name, sub = sbf_get_sub(), main = sbf_get_main())

Arguments

x_name

A string of the name.

sub

A string specifying the path to the sub folder (by default the current sub folder).

main

A string specifying the path to the main folder (by default the current main folder)

Value

A flag specifying whether the data exists.

See also