Get Archive Directory

Get Archive Directory

sbf_get_archive(main = sbf_get_main(), archive = 1L)

Arguments

main

A string specifying the path to the main folder (by default the current main folder)

archive

A positive whole number specifying the folder archived folder where 1L (the default) indicates the most recently archived folder or a character string of the path to the archived folder.

Value

A string of the path to the archived directory.

See also