Load Datas

Load Datas

sbf_load_datas(
  sub = sbf_get_sub(),
  main = sbf_get_main(),
  rename = identity,
  env = parent.frame()
)

Arguments

sub

A string specifying the path to the sub folder (by default the current sub folder).

main

A string specifying the path to the main folder (by default the current main folder)

rename

A single function argument which takes a character vector and returns a character vector of the same length. Used to rename objects before they are loaded into the environment.

env

The environment to the objects into

Value

A invisble character vector of the data frames' names.

See also