Load Data Frames as List Column in Data Frame

Recursively loads all the data frames with names matching the regular expression x_name as the first (list) column (named data) in a data frame. Subsequent character vector columns specify the object names (named name) and sub folders (named sub1, sub2 etc).

Usage

sbf_load_datas_recursive(
  x_name = ".*",
  sub = sbf_get_sub(),
  main = sbf_get_main(),
  include_root = TRUE
)

Arguments

x_name

A string of the name.

sub

A string specifying the path to the sub folder (by default the current sub folder).

main

A string specifying the path to the main folder (by default the current main folder)

include_root

A flag indicating whether to include objects in the top sub folder.

See also