Save Data Frames

Save Data Frames

Usage

sbf_save_datas(
  sub = sbf_get_sub(),
  main = sbf_get_main(),
  env = parent.frame()
)

Arguments

sub

A string specifying the path to the sub folder (by default the current sub folder).

main

A string specifying the path to the main folder (by default the current main folder)

env

An environment.

Value

An invisible character vector of the paths to the saved objects.

See also