Save sf data frames to Geopackages

An sf object of file name file_name is saved as file_name.gpkg.

Usage

sbf_save_gpkgs(
  sub = sbf_get_sub(),
  main = sbf_get_main(),
  env = parent.frame(),
  all_sfcs = TRUE
)

Arguments

sub

A string specifying the path to the sub folder (by default the current sub folder).

main

A string specifying the path to the main folder (by default the current main folder)

env

An environment.

all_sfcs

A flag specifying whether to save non-active sfc columns as geopackages.

Value

An invisible character vector of the paths to the saved objects.

Details

By default (all_sfcs = TRUE) non-active sfc columns are saved as file_name_geometry_column_name.gpkg this includes data frames with no active sfc column.

See also