Gets Subs of a Table as a Character Vector

Recursively returns all the subs of table with name x_name.

sbf_subs_table_recursive(
  x_name,
  sub = sbf_get_sub(),
  main = sbf_get_main(),
  include_root = TRUE
)

Arguments

x_name

A string of the name.

sub

A string specifying the path to the sub folder (by default the current sub folder).

main

A string specifying the path to the main folder (by default the current main folder)

include_root

A flag indicating whether to include objects in the top sub folder.

See also